feat: refactor file updating script to be reused for pip dependencies
